chore: extract modal into a seprate package
