+++ example/src/screens/withModalProvider.tsx
@@ -1,5 +1,5 @@
 import React, { FC } from 'react';
-import { BottomSheetModalProvider } from '@gorhom/bottom-sheet';
+import { BottomSheetModalProvider } from '@gorhom/bottom-sheet-modal';
 
 const withModalProvider = (Component: FC) => () => {
   return (
