+++ packages/bottom-sheet-modal/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@gorhom/bottom-sheet-modal",
   "version": "1.0.0",
-  "description": "A performant interactive bottom sheet with fully configurable options 🚀",
+  "description": "A performant interactive bottom sheet modal with fully configurable options 🚀",
   "main": "lib/commonjs/index",
   "module": "lib/module/index",
   "types": "lib/typescript/index.d.ts",
@@ -18,12 +18,13 @@
     "bottomsheet",
     "reanimated",
     "sheet",
-    "modal"
+    "modal",
+    "presentation-modal"
   ],
   "repository": {
     "type": "git",
     "url": "https://github.com/gorhom/react-native-bottom-sheet",
-    "directory": "packages/bottom-sheet"
+    "directory": "packages/bottom-sheet-modal"
   },
   "author": "Mo Gorhom (https://gorhom.dev)",
   "license": "MIT",
@@ -55,6 +56,7 @@
     "typescript": "^4.1.2"
   },
   "peerDependencies": {
+    "@gorhom/bottom-sheet": ">=2.0.0",
     "react": "*",
     "react-native": "*",
     "react-native-gesture-handler": ">=1.6.1",
