+++ packages/bottom-sheet-modal/src/hooks/index.ts
@@ -1,0 +1,1 @@
+export { useBottomSheetModal } from './useBottomSheetModal';
