+++ packages/bottom-sheet-modal/src/index.ts
@@ -1,0 +1,2 @@
+export { useBottomSheetModal } from './hooks';
+export { default as BottomSheetModalProvider } from './components/bottomSheetModalProvider';
