+++ packages/bottom-sheet-modal/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "../../tsconfig",
+  "compilerOptions": {
+    "rootDir": "./src",
+    "outDir": "./lib/typescript"
+  },
+  "include": ["src"]
+}
